requirements.txt
soundfile
huggingface_hub
yt-dlp
moviepy
numpy
pillow
requests
edge-tts
scipy
imageio
imageio-ffmpeg
pexels-api
opencv-python
pyttsx3
colorspacious
selenium>=4.0
beautifulsoup4>=4.12
webdriver-manager
matplotlib
PyQt6
PyQt6-Qt6
tweepy>=4.14.0
chatterbox-tts
torch
torchaudio
